pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "shoe-tracker"
version = "0.1.0"
description = "Personal running shoe price tracker"
requires-python = ">=3.11"
dependencies = [
    "httpx==0.28.1",
    "selectolax==0.4.7",
    "pydantic==2.13.3",
    "pyyaml==6.0.1",
    "jinja2==3.1.6",
    "click==8.3.2",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.3",
    "ruff==0.15.12",
]

[project.scripts]
shoe-tracker = "shoe_tracker.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
shoe_tracker = ["db/migrations/*.sql", "templates/*.html"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["scripts"]
markers = [
    "live: integration tests that hit real retailer sites (env-gated)",
]

[tool.ruff]
line-length = 110
target-version = "py311"
extend-exclude = ["tests/fixtures"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
]

